+++ src/app/commands/bounty/Bounty.ts
@@ -48,62 +48,43 @@ module.exports = class Bounty extends SlashCommand {
 							description: 'Hash ID of the bounty',
 							required: true,
 						},
-						{
-							name: 'is-complete',
-							type: CommandOptionType.BOOLEAN,
-							description: 'Is the bounty complete as per criteria?',
-							required: true,
-						},
 					],
 				},
 				{
 					name: 'create',
-					type: CommandOptionType.SUB_COMMAND_GROUP,
-					description: 'Create a bounty for the bounty board',
+					type: CommandOptionType.SUB_COMMAND,
+					description: 'Create a new draft of a bounty and finalize on the website',
 					options: [
 						{
-							name: 'new',
-							type: CommandOptionType.SUB_COMMAND,
-							description: 'Create a new draft of a bounty and finalize on the website',
-							options: [
-								{
-									name: 'title',
-									type: CommandOptionType.STRING,
-									description: 'What should the bounty be called?',
-									required: true,
-								},
-								{
-									name: 'summary',
-									type: CommandOptionType.STRING,
-									description: 'What would you like to be worked on?',
-									required: true,
-								},
-								{
-									name: 'criteria',
-									type: CommandOptionType.STRING,
-									description: 'What is absolutely required for this bounty?',
-									required: true,
-								},
-								{
-									name: 'reward',
-									type: CommandOptionType.STRING,
-									description: 'What is the reward? (i.e 100 BANK)',
-									required: true,
-								},
-							],
+							name: 'title',
+							type: CommandOptionType.STRING,
+							description: 'What should the bounty be called?',
+							required: true,
 						},
 						{
-							name: 'publish',
-							type: CommandOptionType.SUB_COMMAND,
-							description: 'Validate discord handle drafted bounty from the website',
-							options: [
-								{
-									name: 'bounty-id',
-									type: CommandOptionType.STRING,
-									description: 'Bounty hash ID',
-									required: true,
-								},
-							],
+							name: 'reward',
+							type: CommandOptionType.STRING,
+							description: 'What is the reward? (i.e 100 BANK)',
+							required: true,
+						},
+						{
+							name: 'copies',
+							type: CommandOptionType.INTEGER,
+							description: 'How many bounties should be published? (level 3+, max 100)',
+							required: false,
+						},
+					],
+				},
+				{
+					name: 'publish',
+					type: CommandOptionType.SUB_COMMAND,
+					description: 'Validate discord handle drafted bounty from the website',
+					options: [
+						{
+							name: 'bounty-id',
+							type: CommandOptionType.STRING,
+							description: 'Bounty hash ID',
+							required: true,
 						},
 					],
 				},
@@ -215,30 +196,29 @@ module.exports = class Bounty extends SlashCommand {
 
 	async run(ctx: CommandContext) {
 		if (ctx.user.bot) return;
-		console.log(`/bounty start ${ctx.user.username}#${ctx.user.discriminator}`);
+		console.log(`start /bounty ${ctx.user.username}#${ctx.user.discriminator}`);
 
 		const { guildMember } = await ServiceUtils.getGuildAndMember(ctx);
 		let command: Promise<any>;
+		let params;
+		
 		switch (ctx.subcommands[0]) {
 		case 'claim':
 			console.log('/bounty claim');
 			command = ClaimBounty(guildMember, ctx.options.claim['bounty-id']);
 			break;
 		case 'create':
-			if (ctx.subcommands[1] === 'new') {
-				const params = this.buildBountyCreateNewParams(ctx.options.create.new);
-				console.log('/bounty create new ' + params);
-				command = CreateNewBounty(guildMember, params, ctx);
-			} else if (ctx.subcommands[1] === 'publish') {
-				console.log('/bounty create publish ');
-				command = PublishBounty(guildMember, ctx.options.create.publish['bounty-id']);
-			} else {
-				return ctx.send(`<@${ctx.user.id}> Sorry command not found, please try again`);
-			}
+			params = this.buildBountyCreateNewParams(ctx.options.create);
+			console.log('/bounty create ' + params.title);
+			command = CreateNewBounty(guildMember, params);
+			break;
+		case 'publish':
+			console.log('/bounty publish ');
+			command = PublishBounty(guildMember, ctx.options.publish['bounty-id']);
 			break;
 		case 'complete':
 			console.log('/bounty complete');
-			command = CompleteBounty(guildMember, ctx.options.complete['bounty-id'], ctx.options.complete['is-complete']);
+			command = CompleteBounty(guildMember, ctx.options.complete['bounty-id']);
 			break;
 		case 'delete':
 			console.log('/bounty delete');
@@ -260,7 +240,7 @@ module.exports = class Bounty extends SlashCommand {
 
 	handleCommandError(ctx: CommandContext, command: Promise<any>) {
 		command.then(() => {
-			console.log(`/bounty end ${ctx.user.username}#${ctx.user.discriminator}`);
+			console.log(`end /bounty ${ctx.user.username}#${ctx.user.discriminator}`);
 			return ctx.send(`${ctx.user.mention} Sent you a DM with information.`);
 		}).catch(e => {
 			console.error('ERROR', e);
@@ -274,17 +254,18 @@ module.exports = class Bounty extends SlashCommand {
 	
 	buildBountyCreateNewParams(ctxOptions): BountyCreateNew {
 		const [reward, symbol] = (ctxOptions.reward != null) ? ctxOptions.reward.split(' ') : [null, null];
+		const copies = (ctxOptions.copies == null || ctxOptions.copies <= 0) ? 1 : ctxOptions.copies;
 		let scale = reward.split('.')[1]?.length;
 		scale = (scale != null) ? scale : 0;
 		return {
 			title: ctxOptions.title,
-			summary: ctxOptions.summary,
-			criteria: ctxOptions.criteria,
 			reward: {
-				amount: reward.replace('.', ''),
+				amount: reward,
 				currencySymbol: symbol,
 				scale: scale,
+				amountWithoutScale:  reward.replace('.', ''),
 			},
+			copies: copies,
 		};
 	}
 };