+++ src/app/service/bounty/RefreshBounty.ts
@@ -37,26 +37,49 @@ export default async (guildMember: GuildMember, bountyId: string, message: Messa
 		embedMessage.setColor('#1e7e34');
 		embedMessage.setDescription(bountyCollection.description);
 		embedMessage.setFooter('🏴 - start | 🔄 - refresh | 📝 - edit | ❌ - delete');
-		embedMessage.fields[2].value = BountyUtils.formatBountyAmount(bountyCollection.reward.amount as number, bountyCollection.reward.scale as number) + ' ' + bountyCollection.reward.currency;
-		embedMessage.fields[1].value = bountyCollection.criteria;
+		embedMessage.fields[2] = {
+			name: 'Reward',
+			value: BountyUtils.formatBountyAmount(bountyCollection.reward.amount as number, bountyCollection.reward.scale as number) + ' ' + bountyCollection.reward.currency,
+			inline: true,
+		};
+		embedMessage.fields[1] = {
+			name: 'Criteria',
+			value: bountyCollection.criteria,
+			inline: false,
+		};
 		await message.edit({ embeds: [embedMessage] });
 		addPublishReactions(message);
 		break;
 	case 'In-Progress':
 		embedMessage.setColor('#d39e00');
 		embedMessage.setFooter('📮 - submit | 🔄 - refresh | 🆘 - help');
+		embedMessage.fields[6] = {
+			name: 'Claimed by',
+			value: bountyCollection.claimedBy.discordHandle,
+			inline: true,
+		};
 		await message.edit({ embeds: [embedMessage] });
 		addClaimReactions(message);
 		break;
 	case 'In-Review':
 		embedMessage.setColor('#d39e00');
 		embedMessage.setFooter('✅ - complete | 🔄 - refresh | 🆘 - help');
+		embedMessage.fields[7] = {
+			name: 'Submitted by',
+			value: bountyCollection.submittedBy.discordHandle,
+			inline: true,
+		};
 		await message.edit({ embeds: [embedMessage] });
 		addSubmitReactions(message);
 		break;
 	case 'Completed':
 		embedMessage.setColor('#1d2124');
 		embedMessage.setFooter('🆘 - help');
+		embedMessage.fields[8] = {
+			name: 'Reviewed by',
+			value: bountyCollection.reviewedBy.discordHandle,
+			inline: true,
+		};
 		await message.edit({ embeds: [embedMessage] });
 		addCompletedReactions(message);
 		break;