+++ src/app/types/bounty/BountyCreateNew.ts
@@ -2,7 +2,9 @@ import { BountyReward } from './BountyReward';
 
 export type BountyCreateNew = {
 	title: string,
-	summary: string,
-	criteria: string,
-	reward: BountyReward
+	summary?: string,
+	criteria?: string,
+	reward: BountyReward,
+	dueAt?: Date,
+	copies: number,
 };